separation of logic
